refactor: move deployment tool to lib/sanity/plugins/tools/ directory
Let's organize plugins by type for convention
